exercises/lab_01_deep_research/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deep_research"]

[project]
name = "deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.2",
    "ddgs>=9.10.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.4",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "tabulate>=0.9.0",
]

[project.scripts]
ui = "deep_research.ui:main"

[dependency-groups]
dev = [
    "ipython>=9.8.0",
]

[tool.uv]
package = true

[tool.pyrefly]
